prune away a lot of cruft related to the old hard-coded uint32_t seeding

+++ MurmurHash3.cpp
@@ -253,7 +253,7 @@ void MurmurHash3_x86_128 ( const void * key, const int len,
 //-----------------------------------------------------------------------------
 
 void MurmurHash3_x64_128 ( const void * key, const int len,
-                           const uint32_t seed, void * out )
+                           const uint64_t seed, void * out )
 {
   const uint8_t * data = (const uint8_t*)key;
   const int nblocks = len / 16;
@@ -346,7 +346,7 @@ void MurmurHash3_x86_128_with_state ( const void * key, int len,
 void MurmurHash3_x64_128_with_state ( const void * key, int len,
                           const void *state, void * out )
 {
-    MurmurHash3_x64_128(key, len, *((uint32_t *)state), out);
+    MurmurHash3_x64_128(key, len, *((uint64_t *)state), out);
 }
 
 //-----------------------------------------------------------------------------
